% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hurricane.R
\name{geom_hurricane}
\alias{geom_hurricane}
\title{geom_hurricane}
\usage{
\dontrun {
map_plot <- ggmap::get_map("Lousiana", zoom = 5, maptype = "toner-background")
map_plot %>%
  ggmap::ggmap(extent = "device") +
  geom_hurricane(data = storm_observation_ike,
                 ggplot2::aes(x = longitude, y = latitude,
                              r_ne = ne, r_se = se, r_nw = nw, r_sw = sw,
                              color = wind_speed, fill = wind_speed)) +
  ggplot2::scale_color_manual(name = "Wind speed (kts)",
                              values = c("red", "orange", "yellow")) +
  ggplot2::scale_fill_manual(name = "Wind speed (kts)",
                             values = c("red", "orange", "yellow"))
}
}
\description{
a wind radii geom that shows a polygon for the three levels of the wind according to wind radii provided in the storm observation
}
\section{aesthetics }{

\itemize{
  \item \strong{x} : the longitude of the center of the storm
  \item \strong{y} : the latitude of the center of the storm
  \item \strong{r_ne} : the wind radii in the northeast
  \item \strong{r_se} : the wind radii in the southeast
  \item \strong{r_nw} : the wind radii in the northwest
  \item \strong{r_sw} : the wind radii in the southwest
  \item colour
  \item fill
  \item size
  \item linetype
  \item alpha
  \item \strong{scale_radii} the maximum radii value
}
}

